Revert "getting chest rendering to work properly (lid animation)"
This reverts commit c0d01c25308144d5b946ccd4a83834237b5995f9.
