+++ src/main/java/com/someguyssoftware/treasure2/block/AbstractChestBlock.java
@@ -5,14 +5,12 @@ package com.someguyssoftware.treasure2.block;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.function.Supplier;
 
 import javax.annotation.Nullable;
 
 import com.someguyssoftware.gottschcore.block.ModContainerBlock;
-import com.someguyssoftware.gottschcore.spatial.Coords;
-import com.someguyssoftware.gottschcore.spatial.Heading;
 import com.someguyssoftware.gottschcore.spatial.Rotate;
-import com.someguyssoftware.gottschcore.world.WorldInfo;
 import com.someguyssoftware.treasure2.Treasure;
 import com.someguyssoftware.treasure2.chest.ILockSlot;
 import com.someguyssoftware.treasure2.chest.TreasureChestType;
@@ -23,33 +21,21 @@ import com.someguyssoftware.treasure2.tileentity.AbstractTreasureChestTileEntity
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.inventory.IInventory;
-import net.minecraft.inventory.InventoryHelper;
-import net.minecraft.inventory.container.INamedContainerProvider;
 import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.pathfinding.PathType;
 import net.minecraft.state.EnumProperty;
 import net.minecraft.state.StateContainer;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ActionResultType;
+import net.minecraft.tileentity.TileEntityType;
 import net.minecraft.util.Direction;
-import net.minecraft.util.Hand;
 import net.minecraft.util.Mirror;
 import net.minecraft.util.Rotation;
+import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
-import net.minecraftforge.fml.network.NetworkHooks;
 
 /**
  * @author Mark Gottschling on Sep 16, 2018
@@ -59,10 +45,16 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	public static final EnumProperty<Direction> FACING = EnumProperty.create("facing", Direction.class);
 
 	/*
+	 * OLD way
 	 *  the class of the tileEntityClass this BlockChest should use.
 	 */
 	private Class<?> tileEntityClass;
 
+	/**
+	 * NEW way
+	 */
+//	protected final TileEntityType<? extends E> tileEntityType;
+	
 	/*
 	 * the concrete object of the tile entity
 	 */
@@ -79,23 +71,23 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	private Rarity rarity;
 
 	/*
-	 * An array of VoxelShape bounds for the bounding box
+	 * An array of AxisAlignedBB bounds for the bounding box
 	 */
-	private VoxelShape[] bounds = new VoxelShape[4];
+//	private AxisAlignedBB[] bounds = new AxisAlignedBB[4];
 
+	private VoxelShape[] bounds = new VoxelShape[4];
+	
 	/**
 	 * 
 	 */
-	public AbstractChestBlock(String modID, String name, Class<? extends AbstractTreasureChestTileEntity> te, 
-			TreasureChestType type, Rarity rarity, Block.Properties properties) {
-		
+	public AbstractChestBlock(String modID, String name, Class<? extends AbstractTreasureChestTileEntity> te, TreasureChestType type, Rarity rarity, Block.Properties properties) {
 		super(modID, name, properties);
 		setTileEntityClass(te);
 		setChestType(type);
 		setRarity(rarity);
 
-		// set the default bounds/shape
 		VoxelShape shape = Block.makeCuboidShape(1.0D, 0.0D, 1.0D, 14.0D, 14.0D, 14.0D);
+		// set the default bounds
 		setBounds(
 				new VoxelShape[] {
 						shape, 	// N
@@ -144,218 +136,23 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
 		builder.add(FACING);
 	}
-
+	
 	/**
 	 * 
 	 */
 	@Override
 	public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
 		switch(state.get(FACING)) {
-		default:
-		case NORTH:
-			return bounds[0];
-		case EAST:
-			return bounds[1];
-		case SOUTH:
-			return bounds[2];
-		case WEST:
-			return bounds[3];
+			default:
+			case NORTH:
+				return bounds[0];
+			case EAST:
+				return bounds[1];
+			case SOUTH:
+				return bounds[2];
+			case WEST:
+				return bounds[3];
 		}
-	}
-
-	@Override
-	public BlockState getStateForPlacement(BlockItemUseContext context) {
-		BlockState blockState = this.getDefaultState().with(FACING,
-				context.getPlacementHorizontalFacing().getOpposite());
-		return blockState;
-	}
-
-	/**
-	 * 
-	 */
-	@Override
-	public boolean isNormalCube(BlockState state, IBlockReader world, BlockPos pos) {
-		return false;
-	}
-	
-	/**
-	 * Called just after the player places a block.
-	 */
-	@Override
-	public void onBlockPlacedBy(World worldIn, BlockPos pos, BlockState state, LivingEntity placer, ItemStack stack) {
-		Treasure.LOGGER.debug("Placing chest from item");
-
-		boolean shouldRotate = false;
-		boolean shouldUpdate = false;
-		boolean forceUpdate = false;
-		AbstractTreasureChestTileEntity tcte = null;
-		Heading oldPersistedChestDirection = Heading.NORTH;
-
-		// face the block towards the player (there isn't really a front)
-		worldIn.setBlockState(pos, state.with(FACING, placer.getHorizontalFacing().getOpposite()), 3);
-		TileEntity te = worldIn.getTileEntity(pos);
-		if (te != null && te instanceof AbstractTreasureChestTileEntity) {
-			// get the backing tile entity
-			tcte = (AbstractTreasureChestTileEntity) te;
-
-			// set the name of the chest
-			if (stack.hasDisplayName()) {
-				tcte.setCustomName(stack.getDisplayName());
-			}
-
-			// read in nbt
-			if (stack.hasTag()) {
-				tcte.readFromItemStackNBT(stack.getTag());
-				forceUpdate = true;
-
-				// get the old tcte facing direction
-				oldPersistedChestDirection = Heading.fromDirection(tcte.getFacing()); // TODO might be byHorizontalIndex
-
-				// dump stack NBT
-//				if (Treasure.LOGGER.isDebugEnabled()) {
-//					dump(stack.getTag(), new Coords(pos), "STACK ITEM -> CHEST NBT");
-//				}
-			}
-
-			// get the direction the block is facing.
-			Heading direction = Heading.fromDirection(placer.getHorizontalFacing().getOpposite());
-
-			// rotate the lock states
-			shouldUpdate = rotateLockStates(worldIn, pos, oldPersistedChestDirection.getRotation(direction)); // old ->
-			// Direction.NORTH
-			// //
-
-						Treasure.LOGGER.debug("New lock states ->");
-						for (LockState ls : tcte.getLockStates()) {
-							Treasure.LOGGER.debug(ls);
-						}
-
-			// update the TCTE facing
-			tcte.setFacing(placer.getHorizontalFacing().getOpposite().getIndex());
-		}
-		if ((forceUpdate || shouldUpdate) && tcte != null) {
-			// update the client
-			tcte.sendUpdates();
-		}
-	}
-
-	/**
-	 * 
-	 */
-	@Override
-	public ActionResultType onBlockActivated(BlockState state, World world, BlockPos pos, PlayerEntity player,
-			Hand handIn, BlockRayTraceResult hit) {
-
-		AbstractTreasureChestTileEntity tileEntity = (AbstractTreasureChestTileEntity) world.getTileEntity(pos);
-
-		// exit if on the client
-		if (WorldInfo.isClientSide(world)) {
-			return ActionResultType.SUCCESS;
-		}
-
-		boolean isLocked = false;
-		// determine if chest is locked
-		if (!tileEntity.hasLocks()) {
-			// get the container provider
-			INamedContainerProvider namedContainerProvider = this.getContainer(state, world, pos);			
-			// open the chest
-			NetworkHooks.openGui((ServerPlayerEntity)player, namedContainerProvider, (packetBuffer)->{});
-			// NOTE: (packetBuffer)->{} is just a do-nothing because we have no extra data to send
-		}
-
-		return ActionResultType.SUCCESS;
-	}
-
-
-	@Override
-	public void onPlayerDestroy(IWorld worldIn, BlockPos pos, BlockState state) {
-		Treasure.LOGGER.info("block destroyed by player. should happen after block is broken/replaced");
-		super.onPlayerDestroy(worldIn, pos, state);
-	}
-
-
-	@Override
-	public void onReplaced(BlockState state, World worldIn, BlockPos pos, BlockState newState, boolean isMoving) {
-		Treasure.LOGGER.debug("Breaking block....!");
-
-		TileEntity tileEntity = worldIn.getTileEntity(pos);
-		AbstractTreasureChestTileEntity te = null;
-		if (tileEntity instanceof AbstractTreasureChestTileEntity) {
-			te = (AbstractTreasureChestTileEntity)tileEntity;
-		}
-
-		if (te != null) {
-			// unlocked!
-			if (!te.hasLocks()) {
-				/*
-				 * spawn inventory items
-				 */
-				InventoryHelper.dropInventoryItems(worldIn, pos, (IInventory) te);
-
-				/*
-				 * spawn chest item
-				 */
-				ItemStack chestItem = new ItemStack(Item.getItemFromBlock(this), 1);
-				Treasure.LOGGER.debug("Item being created from chest -> {}", chestItem.getItem().getRegistryName());
-				InventoryHelper.spawnItemStack(worldIn, (double) pos.getX(), (double) pos.getY(), (double) pos.getZ(),
-						chestItem);
-
-				/*
-				 * write the properties to the nbt
-				 */
-				if (!chestItem.hasTag()) {
-					chestItem.setTag(new CompoundNBT());
-				}
-				te.writePropertiesToNBT(chestItem.getTag());
-			} else {
-				Treasure.LOGGER.debug("[BreakingBlock] ChestConfig is locked, save locks and items to NBT");
-
-				/*
-				 * spawn chest item
-				 */
-
-				if (WorldInfo.isServerSide(worldIn)) {
-					ItemStack chestItem = new ItemStack(Item.getItemFromBlock(this), 1);
-
-					// give the chest a tag compound
-					//					Treasure.LOGGER.debug("[BreakingBlock]Saving chest items:");
-
-					CompoundNBT nbt = new CompoundNBT();
-					nbt = te.write(nbt);
-					chestItem.setTag(nbt);
-
-					InventoryHelper.spawnItemStack(worldIn, (double) pos.getX(), (double) pos.getY(),
-							(double) pos.getZ(), chestItem);
-
-					// TEST log all items in item
-					//					NonNullList<ItemStack> items = NonNullList.<ItemStack>withSize(27, ItemStack.EMPTY);
-					//					ItemStackHelper.loadAllItems(chestItem.getTagCompound(), items);
-					//					for (ItemStack stack : items) {
-					//						Treasure.LOGGER.debug("[BreakingBlock] item in chest item -> {}", stack.getDisplayName());
-					//					}
-				}
-			}
-
-			// remove the tile entity
-			worldIn.removeTileEntity(pos);
-		}
-		else {
-			// default to regular block break;
-			//			super.breakBlock(worldIn, pos, state);
-			//			worldIn.destroyBlock()
-			//			super.onBlockHarvested(worldIn, pos, state, player);
-			super.onReplaced(state, worldIn, pos, newState, isMoving);
-		}
-	}
-
-	// replaces getItemDropped
-	public static void spawnDrops(BlockState state, World worldIn, BlockPos pos) {
-		return;
-	}
-
-	@Override
-	public boolean allowsMovement(BlockState state, IBlockReader worldIn, BlockPos pos, PathType type) {
-		return false;
 	}
 	
 	/**
@@ -376,6 +173,19 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	public static boolean isOpaque(VoxelShape shape) {
 		return false;
 	}
+
+	// used by the renderer to control lighting and visibility of other blocks, also by
+	// (eg) wall or fence to control whether the fence joins itself to this block
+	// set to false because this block doesn't fill the entire 1x1x1 space
+	//	@Override
+	//	public boolean isFullCube(BlockState state) {
+	//		return false;
+	//	}
+
+	//	@Override
+	//	public boolean isNormalCube(BlockState state, IBlockAccess world, BlockPos pos) {
+	//		return false;
+	//	}
 
 	/**
 	 * 
@@ -432,6 +242,14 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	}
 
 	/**
+	 * 
+	 */
+	@Nullable
+	public BlockState getStateForPlacement(BlockItemUseContext context) {
+		return this.getDefaultState().with(FACING, context.getPlayer().getHorizontalFacing().getOpposite());
+	}
+
+	/**
 	 * Returns the blockstate with the given rotation from the passed blockstate. If inapplicable, returns the passed
 	 * blockstate.
 	 * @deprecated call via {@link IBlockState#withRotation(Rotation)} whenever possible. Implementing/overriding is
@@ -450,6 +268,11 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	@Override
 	public BlockState mirror(BlockState state, Mirror mirrorIn) {
 		return state.rotate(mirrorIn.toRotation(state.get(FACING)));
+	}
+
+	@Override
+	public boolean allowsMovement(BlockState state, IBlockReader worldIn, BlockPos pos, PathType type) {
+		return false;
 	}
 
 	/**
@@ -476,7 +299,7 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	/**
 	 * @param tileEntity the tileEntity to set
 	 */
-	public AbstractChestBlock<E> setTileEntity(AbstractTreasureChestTileEntity tileEntity) {
+	public AbstractChestBlock setTileEntity(AbstractTreasureChestTileEntity tileEntity) {
 		this.tileEntity = tileEntity;
 		return this;
 	}
@@ -491,7 +314,7 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	/**
 	 * @param chestType the chestType to set
 	 */
-	public AbstractChestBlock<E> setChestType(TreasureChestType chestType) {
+	public AbstractChestBlock setChestType(TreasureChestType chestType) {
 		this.chestType = chestType;
 		return this;
 	}
@@ -506,17 +329,33 @@ public abstract class AbstractChestBlock<E extends TileEntity> extends ModContai
 	/**
 	 * @param rarity the rarity to set
 	 */
-	public AbstractChestBlock<E> setRarity(Rarity rarity) {
+	public AbstractChestBlock setRarity(Rarity rarity) {
 		this.rarity = rarity;
 		return this;
 	}
 
+	/**
+	 * @return the bounds
+	 */
+//	public AxisAlignedBB[] getBounds() {
+//		return bounds;
+//	}
+
 	public VoxelShape[] getBounds() {
 		return bounds;
 	}
-
-	public AbstractChestBlock<E> setBounds(VoxelShape[] bounds) {
+	
+	/**
+	 * @param bounds the bounds to set
+	 */
+//	public AbstractChestBlock  setBounds(AxisAlignedBB[] bounds) {
+//		this.bounds = bounds;
+//		return this;
+//	}
+	
+	public AbstractChestBlock setBounds(VoxelShape[] bounds) {
 		this.bounds = bounds;
 		return this;
 	}
+
 }
